Enable TS build, implement server in TS.
