+++ packages/client/package.json
@@ -24,7 +24,7 @@
   "scripts": {
     "clean": "rimraf lib/* coverage/*",
     "prepare": "npm run build",
-    "prepublishOnly": "npm run test",
+    "prepublishOnly": "npm run clean && npm run lint && npm run test",
     "build": "tsc",
     "lint:ts": "tsc --noEmit",
     "lint:eslint": "eslint src test",
