+++ packages/client/tsconfig.json
@@ -1,14 +1,9 @@
 {
+  "extends": "../../tsconfig.json",
+
   "compilerOptions": {
     "outDir": "lib",
-    "target": "esnext",
-    "module": "esnext",
-    "skipLibCheck": true,
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "strict": true,
     "removeComments": false,
-    "esModuleInterop": true,
     "declaration": true
   },
   "include": ["src/**/*"]
