--- packages/server/lib/ipfs.js
@@ -1,5 +1,0 @@
-import { create } from "ipfs-http-client";
-import { ipfsAPI } from "./conf.js";
-
-const ipfs = create(ipfsAPI);
-export default ipfs;
