+++ packages/server/package.json
@@ -14,29 +14,37 @@
   "author": "Mathijs de Bruin <mathijs@mathijsfietst.nl>",
   "homepage": "https://github.com/ipfs-search/nyats/tree/main/packages/server#readme",
   "license": "AGPL-3.0-or-later",
-  "main": "lib/cluster.js",
+  "main": "src/cluster.js",
   "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ipfs-search/nyats.git"
   },
   "scripts": {
-    "start": "node lib/cluster.js",
-    "start:watch": "nodemon lib/cluster.js",
+    "start": "ts-node src/cluster.js",
+    "start:watch": "nodemon src/cluster.js",
+    "clean": "rimraf lib/* coverage/*",
+    "prepare": "npm run build",
+    "prepublishOnly": "npm run clean && npm run lint && npm run test",
+    "build": "tsc",
+    "lint:ts": "tsc --noEmit",
+    "lint:eslint": "eslint src test",
+    "lint": "npm run lint:ts && npm run lint:eslint",
     "test": "mocha",
     "test:watch": "nodemon --exec mocha",
-    "test:cover": "c8 --src lib --all mocha",
-    "lint": "eslint lib test"
+    "test:cover": "c8 --src src --all mocha"
   },
   "bugs": {
     "url": "https://github.com/ipfs-search/nyats/issues"
   },
   "devDependencies": {
     "@types/chai": "^4.3.3",
+    "@types/chai-as-promised": "^7.1.5",
     "@types/debug": "^4.1.7",
     "@types/express": "^4.17.13",
     "@types/ffmpeg-static": "^3.0.1",
     "@types/micromatch": "^4.0.2",
+    "@types/node": "^16.11.59",
     "@types/sharp": "^0.30.5",
     "@types/sinon": "^10.0.13",
     "@types/sinon-chai": "^3.2.8",
@@ -45,20 +53,24 @@
     "api-contract-validator": "^2.2.8",
     "babel-eslint": "^10.1.0",
     "c8": "^7.12.0",
+    "chai-as-promised": "^7.1.1",
+    "chai-http": "^4.3.0",
     "mocha": "^10.0.0",
     "mocha-suppress-logs": "^0.3.1",
     "serve-handler": "^6.1.3",
     "sinon": "^14.0.0",
-    "supertest": "^6.2.4"
+    "supertest": "^6.2.4",
+    "ts-sinon": "^2.0.2"
   },
   "dependencies": {
     "@types/mocha": "^9.1.1",
-    "async-iterator-to-stream": "^1.2.0",
     "debug": "^4.3.4",
     "express": "^4.18.1",
     "express-healthcheck": "^0.1.0",
+    "express-validator": "^6.14.2",
     "ffmpeg-static": "^5.1.0",
     "ipfs-http-client": "^51.0.1",
+    "is-ipfs": "^6.0.2",
     "micromatch": "^4.0.5",
     "node-abort-controller": "^3.0.1",
     "sharp": "^0.31.0",
