+++ packages/server/src/audio_thumbnailer.ts
@@ -1,0 +1,6 @@
+import makeFfmpegThumbnailer from "./ffmpeg_thumbnailer";
+import { extractFirstFrame } from "./ffmpeg_extractor";
+
+export default function () {
+  return makeFfmpegThumbnailer(extractFirstFrame);
+}
