+++ packages/server/test/video_thumbnailer.spec.js
@@ -3,8 +3,8 @@ import sharp from "sharp";
 import { expect } from "chai";
 import makeDebugger from "debug";
 
-import makeThumbnailer from "../lib/video_thumbnailer.js";
-import { server } from "./util.js";
+import makeThumbnailer from "../src/video_thumbnailer";
+import { server } from "./util";
 
 const debug = makeDebugger("nyats:spec:video_thumbnailer");
 const thumbnailer = makeThumbnailer();
